messaging-service/requirements.txt
asgiref==3.8.1
channels==4.2.2
colorama==0.4.6
Django==5.2.1
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
pytest==8.3.5
python-dotenv==1.1.0
redis==6.2.0
sqlparse==0.5.3
tzdata==2025.2
